requirements.txt
# Python package requirements for the SaveClient project
# Generated by static analysis of the repository's imports.
# NOTE: Hardware/system packages (mpg123, omxplayer, kernel drivers, etc.) are not Python packages
# and must be installed separately via the OS package manager.

# Core API / AI / TTS / STT
elevenlabs
openai
faster-whisper
srt
python-dotenv

# HTTP
requests

# Raspberry Pi / hardware / GPIO / RFID / NeoPixel
RPi.GPIO
gpiozero
spidev
mfrc522
adafruit-blinka
adafruit-circuitpython-neopixel
rpi_ws281x

# Misc
# (srt included above)

# ---------------------------
# System / runtime notes
# ---------------------------
# - On Raspberry Pi, ensure you have enable I2C/SPI and the appropriate OS-level drivers.
# - For NeoPixel control, you may need additional OS packages or kernel modules; on some setups
#   rpi_ws281x must be installed and configured. If you prefer Blinka/CircuitPython, install
#   adafruit-blinka and adafruit-circuitpython-neopixel as above.
# - Audio players referenced in code (mpg123, omxplayer) are external programs; install via apt:
#     sudo apt update; sudo apt install mpg123 omxplayer
# - If you need to pin versions, replace the package names above with e.g. requests==2.31.0
# - If a package name differs for your platform, adapt accordingly.
